requirements.txt
boto3>=1.28.0
tensorflow-macos>=2.15.0
tensorflow-metal>=1.1.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.31.0
